refactor: group all WI-related testcases into a single test file (#1115)
This commit has no functional changes.
