+++ e2e/nomostest/testutils/finalizer.go
@@ -15,32 +15,8 @@
 package testutils
 
 import (
-	"strings"
-
-	"github.com/pkg/errors"
-	"kpt.dev/configsync/pkg/api/configsync/v1beta1"
-	"kpt.dev/configsync/pkg/util/log"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-// ValidateError returns true if the specified errors contain an error
-// with the specified error code and (partial) message.
-func ValidateError(errs []v1beta1.ConfigSyncError, code, message string) error {
-	if len(errs) == 0 {
-		return errors.Errorf("no errors present")
-	}
-	for _, e := range errs {
-		if e.Code == code {
-			if message == "" || strings.Contains(e.ErrorMessage, message) {
-				return nil
-			}
-		}
-	}
-	if message != "" {
-		return errors.Errorf("error %s not present with message %q: %s", code, message, log.AsJSON(errs))
-	}
-	return errors.Errorf("error %s not present: %s", code, log.AsJSON(errs))
-}
 
 // AppendFinalizer adds a finalizer to the object
 func AppendFinalizer(obj client.Object, finalizer string) {
