+++ pkg/reconcilermanager/controllers/reposync_controller.go
@@ -1087,9 +1087,11 @@ func (r *RepoSyncReconciler) mutationsFor(ctx context.Context, rs *v1beta1.RepoS
 		}
 		injectFWICreds := useFWIAuth(auth, r.membership)
 		if injectFWICreds {
-			if err := r.injectFleetWorkloadIdentityCredentials(&d.Spec.Template, gcpSAEmail); err != nil {
-				return err
+			creds, err := BuildFWICredsContent(r.membership.Spec.WorkloadIdentityPool, r.membership.Spec.IdentityProvider, gcpSAEmail)
+			if err != nil {
+				return nil
 			}
+			core.SetAnnotation(&d.Spec.Template, metadata.FleetWorkloadIdentityCredentials, creds)
 		}
 
 		// Add sync-generation label
